requirements.txt
Django==5.1.6
python-dateutil==2.9.0.post0